+++ src/ayame/ayame_client.h
@@ -30,6 +30,8 @@ struct AyameClientConfig {
   std::string room_id;
   std::string client_id;
   std::string signaling_key;
+  std::string video_codec_type;
+  std::string audio_codec_type;
 };
 
 class AyameClient : public std::enable_shared_from_this<AyameClient>,
@@ -60,6 +62,7 @@ class AyameClient : public std::enable_shared_from_this<AyameClient>,
   void ReconnectAfter();
   void OnWatchdogExpired();
   bool ParseURL(URLParts& parts) const;
+  void SetCodecPreferences();
 
  private:
   void DoRead();